remoced SymbExecTest everywhere, framework/analzyer should work
